# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authority Navigation API
  version: 1.2.0
  description: >
    权威综合导航页后端 API。v1.1.0 加入 x-admin-token 认证与写接口限流；
    v1.2.0 改为「登录一次」模型：口令只在 POST /api/login 校验一次并下发
    HttpOnly 会话 Cookie，之后写接口凭 Cookie 直接通过，不再重复校验口令；
    x-admin-token 仍作为脚本 / CI 的兼容入口保留。读接口保持匿名开放。
servers:
  - url: /
paths:
  /health:
    get:
      summary: 健康检查
      security: []
      responses:
        '200':
          description: 服务存活
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]
                  timestamp:
                    type: string
                    format: date-time
  /api/session:
    get:
      summary: 查询当前登录态（匿名，始终 200）
      security: []
      responses:
        '200':
          description: 是否已登录
          content:
            application/json:
              schema:
                type: object
                required: [authenticated]
                properties:
                  authenticated:
                    type: boolean
  /api/login:
    post:
      summary: 用管理员口令换取登录会话（限流）
      description: >
        口令正确时下发 HttpOnly + SameSite=Strict 的 nav_session Cookie，
        有效期内写接口无需再校验口令；口令错误返回与未登录一致的 401。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: 由服务端环境变量 ADMIN_TOKEN 配置的管理员口令
      responses:
        '200':
          description: 登录成功
          headers:
            Set-Cookie:
              schema:
                type: string
                example: nav_session=1750000000000.signature; Path=/; HttpOnly; SameSite=Strict
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
                  expiresAt:
                    type: string
                    format: date-time
        '401':
          description: 口令错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '429':
          description: 尝试速率超限
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Too many requests
  /api/logout:
    post:
      summary: 退出登录，清除会话 Cookie
      security: []
      responses:
        '200':
          description: 已退出
          content:
            application/json:
              schema:
                type: object
                properties:
                  authenticated:
                    type: boolean
  /api/data:
    get:
      summary: 获取用户自定义数据（匿名）
      security: []
      responses:
        '200':
          description: 当前自定义数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavData'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: 保存用户自定义数据（需登录会话或管理员口令，限流）
      description: >
        全量替换语义。认证顺序为：有效登录会话 Cookie 优先，其次兼容
        x-admin-token 请求头。超过速率限制返回 429（Retry-After 头指示重试等待秒数）。
      security:
        - sessionCookie: []
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavDataInput'
      responses:
        '200':
          description: 保存成功，返回清洗后的数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavData'
        '400':
          description: 载荷校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: customLinks must be an object
        '401':
          description: 未登录（无有效会话 Cookie 且未提供正确口令）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '429':
          description: 写操作速率超限
          headers:
            Retry-After:
              schema:
                type: integer
                description: 建议重试等待秒数
            RateLimit:
              schema:
                type: string
                description: IETF draft-8 合并限流头
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Too many requests
        '500':
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nav_session
      description: 登录后下发的会话 Cookie（HttpOnly + SameSite=Strict，有效期 12 小时）
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: 管理员口令，由服务端环境变量 ADMIN_TOKEN 配置（>=16 字符）；用于脚本 / CI 直接调用
  schemas:
    NavData:
      type: object
      properties:
        customLinks:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/LinkItem'
        customCategories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        removedDefaults:
          type: array
          items:
            type: string
            format: uri
        categoryOrder:
          type: array
          items:
            type: string
    NavDataInput:
      type: object
      description: 与 NavData 同构；服务端会做类型校验、XSS 消毒、URL 协议白名单与长度截断
      properties:
        customLinks:
          type: object
        customCategories:
          type: array
          items:
            type: object
        removedDefaults:
          type: array
          items:
            type: string
        categoryOrder:
          type: array
          items:
            type: string
    LinkItem:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
          maxLength: 100
        url:
          type: string
          maxLength: 2048
          description: 仅接受 http/https 协议
        custom:
          type: boolean
    Category:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        icon:
          type: string
          maxLength: 500
        links:
          type: array
          items:
            $ref: '#/components/schemas/LinkItem'
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
  responses:
    ServerError:
      description: 服务内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error